mcp-agent-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.116.1
frozenlist==1.7.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
instructor==1.10.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.12.2
mcp-agent==0.1.8
mdurl==0.1.2
multidict==6.6.3
numpy==2.3.2
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-distro==0.57b0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-anthropic==0.43.1
opentelemetry-instrumentation-openai==0.43.1
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.11
packaging==25.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
